geodjango/requirements.txt
Django==2.2.5
django-leaflet==0.24.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
